+++ src/components/form/index.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="./checkbox/index.d.ts" />
+/// <reference path="./field_search/index.d.ts" />