+++ src/index.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="./services/index.d.ts" />
+/// <reference path="./components/index.d.ts" />